Update apps.json
A suggestion - JumpLink isn't an "app": that is, it doesn't play podcasts. It's closer to a directory, since it links to apps to play podcasts. It's quite similar to Podnews podcast pages, which are marked as a 'directory' here.

It's confusing to label it as an "app" in the generally accepted sense of being an app, I'd argue.